Split context creation into own package
