+++ ffmpeg.go
@@ -14,9 +14,10 @@ import (
 	"image/color"
 	"io"
 	"strings"
-	"sync"
 	"time"
 	"unsafe"
+
+	"github.com/bakape/video/avio"
 )
 
 func init() {
@@ -24,183 +25,18 @@ func init() {
 	C.avcodec_register_all()
 }
 
-var (
-	// IOBufferSize defines the size of the allocated buffer used for allacating
-	// the C response from ffmpeg
-	IOBufferSize = 4096
-
-	// Global map of AVIOHandlers. One handlers struct per format context.
-	// Using ctx pointer address as a key.
-	handlersMap = handlerMap{
-		m: make(map[uintptr]*Handlers),
-	}
-)
-
-// Functions prototypes for custom IO. Implement necessary prototypes and pass
-// instance pointer to NewAVIOContext.
-//
-// E.g.:
-// 	func gridFsReader() ([]byte, int) {
-// 		... implementation ...
-//		return data, length
-//	}
-//
-//	avoictx := NewAVIOContext(ctx, &AVIOHandlers{ReadPacket: gridFsReader})
-type Handlers struct {
-	ReadPacket  func([]byte) (int, error)
-	WritePacket func([]byte) (int, error)
-	Seek        func(int64, int) (int64, error)
-}
-
-type Context struct {
-	AVIOCtx     *C.struct_AVIOContext
-	AVFormatCtx *C.struct_AVFormatContext
-	handlerKey  uintptr
-}
-
-type handlerMap struct {
-	sync.RWMutex
-	m map[uintptr]*Handlers
-}
-
-func (h *handlerMap) Set(k uintptr, handlers *Handlers) {
-	h.Lock()
-	h.m[k] = handlers
-	h.Unlock()
-}
-
-func (h *handlerMap) Delete(k uintptr) {
-	h.Lock()
-	delete(h.m, k)
-	h.Unlock()
-}
-
-func (h *handlerMap) Get(k unsafe.Pointer) *Handlers {
-	h.RLock()
-	handlers, ok := h.m[uintptr(k)]
-	h.RUnlock()
-	if !ok {
-		panic(fmt.Sprintf(
-			"No handlers instance found, according pointer: %v",
-			k,
-		))
-	}
-	return handlers
-}
-
-// AVIOContext constructor. Use it only if You need custom IO behaviour!
-func NewContext(handlers *Handlers) (*Context, error) {
-	ctx := C.avformat_alloc_context()
-	this := &Context{}
-
-	buffer := (*C.uchar)(C.av_malloc(C.size_t(IOBufferSize)))
-
-	if buffer == nil {
-		return nil, errors.New("unable to allocate buffer")
-	}
-
-	// we have to explicitly set it to nil, to force library using default
-	// handlers
-	var ptrRead, ptrWrite, ptrSeek *[0]byte = nil, nil, nil
-
-	if handlers != nil {
-		this.handlerKey = uintptr(unsafe.Pointer(ctx))
-		handlersMap.Set(this.handlerKey, handlers)
-	}
-
-	if handlers.ReadPacket != nil {
-		ptrRead = (*[0]byte)(C.readCallBack)
-	}
-
-	if handlers.WritePacket != nil {
-		ptrWrite = (*[0]byte)(C.writeCallBack)
-	}
-
-	if handlers.Seek != nil {
-		ptrSeek = (*[0]byte)(C.seekCallBack)
-	}
-
-	this.AVIOCtx = C.avio_alloc_context(
-		buffer,
-		C.int(IOBufferSize),
-		0,
-		unsafe.Pointer(ctx),
-		ptrRead,
-		ptrWrite,
-		ptrSeek,
-	)
-	if this.AVIOCtx == nil {
-		return nil, errors.New("unable to initialize avio context")
-	}
-
-	ctx.pb = this.AVIOCtx
-	if this.AVFormatCtx = C.create_context(ctx); ctx == nil {
-		this.Free()
-		return nil, errors.New("failed to initialize AVFormatContext")
-	}
-
-	return this, nil
-}
-
-// Free frees up resources allocated hor handling I/O in C
-func (a *Context) Free() {
-	handlersMap.Delete(a.handlerKey)
-}
-
-//export readCallBack
-func readCallBack(opaque unsafe.Pointer, buf *C.uint8_t, bufSize C.int) C.int {
-	handlers := handlersMap.Get(opaque)
-	if handlers.ReadPacket == nil {
-		panic("No reader handler initialized")
-	}
-	s := (*[1 << 30]byte)(unsafe.Pointer(buf))[:bufSize:bufSize]
-	n, err := handlers.ReadPacket(s)
-	if err != nil {
-		return -1
-	}
-	return C.int(n)
-}
-
-//export writeCallBack
-func writeCallBack(opaque unsafe.Pointer, buf *C.uint8_t, bufSize C.int) C.int {
-	handlers := handlersMap.Get(opaque)
-	if handlers.WritePacket == nil {
-		panic("No writer handler initialized.")
-	}
-
-	n, err := handlers.WritePacket(C.GoBytes(unsafe.Pointer(buf), bufSize))
-	if err != nil {
-		return -1
-	}
-	return C.int(n)
-}
-
-//export seekCallBack
-func seekCallBack(opaque unsafe.Pointer, offset C.int64_t, whence C.int) C.int64_t {
-	handlers := handlersMap.Get(opaque)
-	if handlers.Seek == nil {
-		panic("No seek handler initialized.")
-	}
-
-	n, err := handlers.Seek(int64(offset), int(whence))
-	if err != nil {
-		return -1
-	}
-	return C.int64_t(n)
-}
-
 // Decode uses CGo FFmpeg binding to extract the first video frame
 func Decode(r io.ReadSeeker) (image.Image, error) {
-	avioCtx, err := NewContext(&Handlers{
+	ctx, err := avio.NewContext(&avio.Handlers{
 		ReadPacket: r.Read,
 		Seek:       r.Seek,
 	})
 	if err != nil {
 		return nil, err
 	}
-	defer avioCtx.Free()
+	defer ctx.Free()
 
-	f := C.extract_video_image(avioCtx.AVFormatCtx)
+	f := C.extract_video_image(ctx.AVFormatContext())
 	if f == nil {
 		return nil, errors.New("Failed to Get AVCodecContext")
 	}
@@ -238,16 +74,16 @@ func Decode(r io.ReadSeeker) (image.Image, error) {
 
 // DecodeConfig uses CGo FFmpeg binding to find video config
 func DecodeConfig(r io.ReadSeeker) (image.Config, error) {
-	avioCtx, err := NewContext(&Handlers{
+	ctx, err := avio.NewContext(&avio.Handlers{
 		ReadPacket: r.Read,
 		Seek:       r.Seek,
 	})
 	if err != nil {
 		return image.Config{}, err
 	}
-	defer avioCtx.Free()
+	defer ctx.Free()
 
-	f := C.extract_video(avioCtx.AVFormatCtx)
+	f := C.extract_video(ctx.AVFormatContext())
 	if f == nil {
 		return image.Config{}, errors.New("Failed to decode")
 	}
@@ -271,23 +107,24 @@ func DecodeConfig(r io.ReadSeeker) (image.Config, error) {
 // DecodeAVFormatDetail retrieves contained stream codecs in more verbose
 // representation
 func DecodeAVFormatDetail(r io.ReadSeeker) (audio, video string, err error) {
-	avioCtx, err := NewContext(&Handlers{
+	ctx, err := avio.NewContext(&avio.Handlers{
 		ReadPacket: r.Read,
 		Seek:       r.Seek,
 	})
 	if err != nil {
 		return
 	}
-	defer avioCtx.Free()
+	defer ctx.Free()
 
-	f := C.extract_video(avioCtx.AVFormatCtx)
+	avfc := ctx.AVFormatContext()
+	f := C.extract_video(avfc)
 	if f == nil {
 		err = errors.New("Failed to decode video stream")
 		return
 	}
 	video = C.GoString(f.codec.long_name)
 
-	f = C.extract_audio(avioCtx.AVFormatCtx)
+	f = C.extract_audio(avfc)
 	if f == nil {
 		err = errors.New("Failed to decode audio stream")
 		return
@@ -298,23 +135,24 @@ func DecodeAVFormatDetail(r io.ReadSeeker) (audio, video string, err error) {
 
 // DecodeAVFormat retrieves contained stream codecs
 func DecodeAVFormat(r io.ReadSeeker) (audio, video string, err error) {
-	avioCtx, err := NewContext(&Handlers{
+	ctx, err := avio.NewContext(&avio.Handlers{
 		ReadPacket: r.Read,
 		Seek:       r.Seek,
 	})
 	if err != nil {
 		return
 	}
-	defer avioCtx.Free()
+	defer ctx.Free()
 
-	f := C.extract_video(avioCtx.AVFormatCtx)
+	avfc := ctx.AVFormatContext()
+	f := C.extract_video(avfc)
 	if f == nil {
 		err = errors.New("Failed to decode video stream")
 		return
 	}
 	video = C.GoString(f.codec.name)
 
-	f = C.extract_audio(avioCtx.AVFormatCtx)
+	f = C.extract_audio(avfc)
 	if f == nil {
 		err = errors.New("Failed to decode audio stream")
 		return
@@ -325,14 +163,15 @@ func DecodeAVFormat(r io.ReadSeeker) (audio, video string, err error) {
 
 // DecodeLength returns the length of the video
 func DecodeLength(r io.ReadSeeker) (time.Duration, error) {
-	avioCtx, err := NewContext(&Handlers{
+	ctx, err := avio.NewContext(&avio.Handlers{
 		ReadPacket: r.Read,
 		Seek:       r.Seek,
 	})
 	if err != nil {
 		return 0, err
 	}
-	defer avioCtx.Free()
+	defer ctx.Free()
 
-	return time.Duration(avioCtx.AVFormatCtx.duration * 1000), nil
+	dur := (*C.struct_AVFormatContext)(ctx.AVFormatContext()).duration
+	return time.Duration(dur * 1000), nil
 }
